Add new API to wake esp32 from deep sleep with external trigger
The new API is part of the gpio module and is called `dsleep_wakeup()`.
It uses the esp_sleep module and allows an RTC pin to be configured as a
wakeup trigger from deep sleep. Input is the GPIO pin number and whether
it should wake on HIGH or LOW. See IDF documentation for more info [1].

[1] https://docs.espressif.com/projects/esp-idf/en/latest/api-reference/system/sleep_modes.html#_CPPv228esp_sleep_enable_ext0_wakeup10gpio_num_ti